Improvements to active sessions widget
